extract.sh: Cleanup code and use proper sh
I have no idea what was going through my head when I didn't just loop
over the small set of files that I knew this repo consisted of and then
made one exception for the lone directory.

In the future if more directories are added then they can also be added
to and checked for in the loop.